Preprocessing and extending
Extended SimpleGulpGlob tests to GulpGlob. After defining methods
toPromise and list, 2 of them pass. But also a preprocessing is
necessary: By default, arrays when passed through a Polyton are
considered list of args for the underlying Singletons. But here
we do want arrays of globs to be treated as a single valid arg.