ready option is now a function
Also refactored dest method.

First change allows for passed promise not to be created in case glob
is rejected, though valid, by wrapping GulpGlop. This happens in dev
stage as current keyfunc can't take 'optional' args and
therefore the ready promise breaks it.

Second change is here to allow to pass correct args to GulpGlop on
running dest. Previous scheme to pass whole SimpleGulpGlob objects
breaks the singletonness which is only ensured at the level of the
Polyton